major refactoring, hence the new branch
