+++ README.md
@@ -79,9 +79,14 @@ will restart the service if it's not writing in the logs for more than 40 second
 
 The rpi is isolated most of the time, from the outside world. That's why we sync the time and we send push notifications
 once an hour. This is to make sure the rpi is still alive and kicking. If we need to change the behavior remotely, we can
-create a AWS server that can be checked every time we 'surface' and if so, we download and replace the new config.
-Unless you keep the setup in a remote location.. this is not needed. But it's an idea.
+create an AWS server that can be checked every time we 'surface' and if so, we download and replace the new config.
+Unless you keep the setup in a remote location.. this is not needed. But it's an idea. This would be nice if you also add
+video capabilities and you want it to take a video when you're not around. Depends on the use case.
 
+### Specs
+
+These are the keep alive times. `[0, 3, 6, 9, 15, 18, 21, 24, 27, 30, 36, 39, 42, 45, 48, 54, 57]`. Try not to mess with this.
+There's some other utils. Use them as you see fit. They're still work in progress.
 
 Enjoy! :)
