+++ specs/GoPro_Time-lapse_full_cycle.txt
