--- timelapse.service
@@ -1,31 +1,0 @@
-#!/bin/bash
-
-[Unit]
-Description=GoPro Timelapse Script
-After=network-online.target
-Wants=network-online.target
-
-StartLimitIntervalSec=180
-StartLimitBurst=3
-
-[Service]
-Type=simple
-
-# Load configuration and execute script
-ExecStart=/bin/bash -c "config_file='/home/timelapse/config.json'; username=$(jq -r .rpi_service.username $config_file); script_path=$(jq -r .rpi_service.path $config_file); work_dir=$(jq -r .rpi_service.work_dir $config_file); . /home/gopro_env/bin/activate && python $script_path --debug"
-
-Restart=on-failure
-RestartSec=60
-#WatchdogSec=120
-
-User=root
-Environment="PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
-WorkingDirectory=/home/timelapse
-
-# ExecStartPre=/bin/bash -c "for i in {1..6}; do ping -c1 8.8.8.8 && exit 0; sleep 4; done; exit 0"
-
-#OnFailure=timelapse_failure.service
-
-[Install]
-WantedBy=multi-user.target
-
