--- timelapse_failure.service
@@ -1,21 +1,0 @@
-[Unit]
-Description=GoPro Timelapse Failure Handler
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=oneshot
-ExecStart=/bin/bash -c "config_file='/home/timelapse/config.json'; \
-    api_key=\$(jq -r .pushbullet.api_key \"\$config_file\"); \
-    message='Time-lapse script cannot be started. Multiple failures. Assistance needed. Will shut-down now..'; \
-    if ping -c 1 8.8.8.8 > /dev/null; then \
-        curl -X POST https://api.pushbullet.com/v2/pushes \
-            -H \"Access-Token: \$api_key\" \
-            -H \"Content-Type: application/json\" \
-            -d '{\"type\":\"note\",\"title\":\"Critical Alert\",\"body\":\"'\"\$(date +'%Y-%m-%d %H:%M:%S') -- \$message\"'}'; \
-    fi; \
-    echo \"This would have shut down the whole rpi.\"; \
-    echo blah "
-
-[Install]
-WantedBy=multi-user.target
