+++ util/env_setup.sh
@@ -1,11 +1,6 @@
 #!/usr/bin/env bash
 #
-# setup_timelapse.sh
-# Prepares a fresh Raspberry Pi OS for the GoPro timelapse script + systemd services.
-#
-# Usage:
-#   chmod +x setup_timelapse.sh
-#   ./setup_timelapse.sh
+# Prepares a fresh Raspberry Pi OS for the GoPro timelapse script + systemd service. (hopefully)
 #
 
 if [ "$(uname)" == "Darwin" ]; then
@@ -17,7 +12,7 @@ set -e  # Exit on any error
 
 TIMELAPSE_DIR="/home/timelapse"
 GOPRO_ENV="/home/gopro_env"
-SVC_USER="mrbigheart"  # Adjust to the user you prefer (e.g. 'pi' or 'mrbigheart')
+SVC_USER="mrbigheart"
 
 echo "==== Updating apt-get and upgrading existing packages ===="
 sudo apt-get update
@@ -50,14 +45,14 @@ echo "==== Activating the venv and installing packages inside it ===="
 sudo -u "$SVC_USER" bash -c "
     source $GOPRO_ENV/bin/activate
     pip install --upgrade pip
-    pip install goprocam watchdog requests pushbullet.py
+    pip install goprocam requests pushbullet.py
 "
 
 echo "==== Upgrading pip ===="
 sudo pip3 install --upgrade pip --break-system-packages
 
 echo "==== Installing Python libraries system-wide (optional) ===="
-sudo pip3 install goprocam watchdog requests pushbullet.py jq --break-system-packages
+sudo pip3 install goprocam requests pushbullet.py jq --break-system-packages
 
 echo "==== Ensuring $TIMELAPSE_DIR exists and is owned by $SVC_USER ===="
 sudo mkdir -p "$TIMELAPSE_DIR"
@@ -78,47 +73,15 @@ StartLimitBurst=3
 [Service]
 Type=simple
 
-ExecStart=/bin/bash -c "config_file='$TIMELAPSE_DIR/config.json'; \
-username=\$(jq -r .rpi_service.username \"\$config_file\"); \
-script_path=\$(jq -r .rpi_service.path \"\$config_file\"); \
-work_dir=\$(jq -r .rpi_service.work_dir \"\$config_file\"); \
-source $GOPRO_ENV/bin/activate && python \"\$script_path\""
+# Load configuration and execute script
+ExecStart=/bin/bash -c "config_file='/home/timelapse/config/config.json'; username=$(jq -r .rpi_service.username $config_file); script_path=$(jq -r .rpi_service.path $config_file); work_dir=$(jq -r .rpi_service.work_dir $config_file); . /home/gopro_env/bin/activate && python $script_path"
 
 Restart=on-failure
 RestartSec=60
-WatchdogSec=120
 
-User=$SVC_USER
-WorkingDirectory=$TIMELAPSE_DIR
-
-ExecStartPre=/bin/bash -c "until ping -c1 8.8.8.8; do sleep 5; done"
-
-OnFailure=timelapse_failure.service
-
-[Install]
-WantedBy=multi-user.target
-EOF
-
-# timelapse_failure.service
-sudo tee /etc/systemd/system/timelapse_failure.service >/dev/null <<'EOF'
-[Unit]
-Description=GoPro Timelapse Failure Handler
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=oneshot
-
-ExecStart=/bin/bash -c "config_file='/home/timelapse/config.json'; \
-api_key=\$(jq -r .pushbullet.api_key \"\$config_file\"); \
-message='Time-lapse script cannot be started. Multiple failures. Assistance needed. Will shut-down now..'; \
-if ping -c 1 8.8.8.8 > /dev/null; then \
-    curl -X POST https://api.pushbullet.com/v2/pushes \
-        -H \"Access-Token: \$api_key\" \
-        -H \"Content-Type: application/json\" \
-        -d '{\"type\":\"note\",\"title\":\"Critical Alert\",\"body\":\"'\"\$(date +'%Y-%m-%d %H:%M:%S') -- \$message\"'}'; \
-fi; \
-echo \"This would have shut down the rpi.\""
+User=root
+Environment="PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+WorkingDirectory=/home/timelapse
 
 [Install]
 WantedBy=multi-user.target
@@ -127,18 +90,7 @@ EOF
 echo "==== Reloading systemd, enabling and starting timelapse services ===="
 sudo systemctl daemon-reload
 sudo systemctl enable timelapse.service
-sudo systemctl enable timelapse_failure.service
 
 echo "==== Setup complete! ===="
 echo "You can now reboot, and the timelapse script will run automatically."
-echo "  sudo reboot"
-
-###############################################################################
-# OPTIONAL: Auto-open a terminal with 'less' on the latest log file after boot.
-# You can do this via .bashrc or systemd as well. For instance:
-#
-# echo "sleep 10 && x-terminal-emulator -e 'less +F /home/timelapse/logs/daily_logs_$(date +%Y_%m_%d).txt'" \
-#     >> /home/$SVC_USER/.bashrc
-#
-# That means after you log in, it waits 10s, then opens a terminal reading the log file.
-###############################################################################
+echo "echoing.. sudo reboot"
