+++ util/monitor_heartbeat.sh
@@ -5,19 +5,17 @@
 
 LOGS_DIR="/home/timelapse/logs"
 SERVICE_NAME="timelapse.service"
-MAX_INTERVAL=40  # Maximum allowed age of the log file in seconds
+MAX_INTERVAL=40
 
 # Find the most recent log file matching the pattern
 LOG_FILE=$(find "$LOGS_DIR" -type f -name "daily_logs_*" -printf '%T@ %p\n' | sort -n | awk 'END {print $2}')
 
-# Check if a log file was found
 if [[ -z "$LOG_FILE" ]]; then
     echo "$(date): No log file found in $LOGS_DIR. Restarting $SERVICE_NAME."
     sudo systemctl restart $SERVICE_NAME
     exit 1
 fi
 
-# Check the file's last modification time
 LAST_MODIFIED=$(stat -c %Y "$LOG_FILE")
 CURRENT_TIME=$(date +%s)
 AGE=$((CURRENT_TIME - LAST_MODIFIED))
@@ -28,4 +26,4 @@ if [[ $AGE -gt $MAX_INTERVAL ]]; then
     exit 1
 fi
 
-echo "$(date): Log file is healthy (age: $AGE seconds, file: $LOG_FILE)."
+echo "$(date): Still kickin' age: $AGE seconds for $LOG_FILE)."
