+++ util/timelapse.service
@@ -1,0 +1,26 @@
+#!/bin/bash
+
+[Unit]
+Description=GoPro Timelapse Script
+After=network-online.target
+Wants=network-online.target
+
+StartLimitIntervalSec=180
+StartLimitBurst=3
+
+[Service]
+Type=simple
+
+# Load configuration and execute script
+ExecStart=/bin/bash -c "config_file='/home/timelapse/config/config.json'; username=$(jq -r .rpi_service.username $config_file); script_path=$(jq -r .rpi_service.path $config_file); work_dir=$(jq -r .rpi_service.work_dir $config_file); . /home/gopro_env/bin/activate && python $script_path"
+
+Restart=on-failure
+RestartSec=60
+
+# root is easier, trust me
+User=root
+Environment="PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+WorkingDirectory=/home/timelapse
+
+[Install]
+WantedBy=multi-user.target
